Source Code/frontend/requirements.txt
flask
flask-cors
torch
torchvision
numpy
pillow
scikit-learn
joblib
